% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rttest.R
\name{myttest}
\alias{myttest}
\title{Title}
\usage{
myttest(x, y, alpha, paired)
}
\arguments{
\item{x}{vector}

\item{y}{vector}

\item{alpha}{t.test parameter}

\item{paired}{boolean}
}
\value{
Rttest object
}
\description{
Title
}
\examples{
\dontrun{myttest(x,y,alpha,TRUE)}
}
