% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.Rttest}
\alias{plot.Rttest}
\title{Boxplot}
\usage{
\method{plot}{Rttest}(x, ...)
}
\arguments{
\item{x}{an object of class Rttest}

\item{...}{other parameters}
}
\value{
a boxplot
}
\description{
Boxplot
}
