% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.Rttest}
\alias{print.Rttest}
\title{Title}
\usage{
\method{print}{Rttest}(x, ...)
}
\arguments{
\item{x}{object Rttest}

\item{...}{other parameters}
}
\value{
. nothing
}
\description{
Title
}
\examples{
\dontrun{print(x)}
}
